Move Libraries to live under CompilerOpts
